commit bug fixed
